added date and time picker UI and made the calendar row of the edit section stop clicking out like the Box around it

+++ app/src/main/java/com/happymeerkat/avocado/domain/use_case/CategoryDelete.kt
@@ -9,7 +9,9 @@ class CategoryDelete(
     private val listRepo: ListRepository
 ) {
     suspend operator fun invoke(category: Category) {
-        catRepo.deleteCategory(category)
-        listRepo.deleteAllByCategory(category)
+        if(category.name != "All") {
+            catRepo.deleteCategory(category)
+            listRepo.deleteAllByCategory(category)
+        }
     }
 }
